feat: refactor useEffect on search component
